pyproject.toml
[project]
name = "qrcode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "ipykernel>=7.1.0",
    "loguru>=0.7.3",
    "omegaconf>=2.3.0",
    "pillow>=12.1.0",
    "python-docx>=1.2.0",
]

[dependency-groups]
dev = [
    "qrcode[pil]",
]

[tool.uv.sources]
qrcode = { workspace = true }
